+++ bycon/config.py
@@ -6,11 +6,12 @@ import socket
 pkg_path = path.dirname( path.abspath(__file__) )
 
 """
-Global variables
-Potentially in the environment:
+Runtime global variables that might be modified through providing them
+in the environment:
 
 DATABASE_NAMES
-DB_MONGOHOST
+BYCON_MONGO_HOST
+BYC_LOCAL_CONF ==> LOC_PATH
 
 """
 
@@ -18,37 +19,44 @@ ENV = environ.get('HTTP_HOST', "___shell___")
 HOSTNAME = environ.get('HOSTNAME', socket.gethostname())
 
 PKG_PATH = pkg_path
-CONF_PATH = path.join( pkg_path, "config")
-LIB_PATH = path.join( pkg_path, "lib")
+CONF_PATH = path.join(pkg_path, "config")
+LIB_PATH = path.join(pkg_path, "lib")
 
 NO_PARAM_VALUES = ["none", "null", "undefined"]
 
 # path of the calling script is used to point to a local config directory
 CALLER_PATH = path.dirname( path.abspath(sys.argv[0]))
 PROJECT_PATH = path.join(CALLER_PATH, pardir)
-LOC_PATH = path.join(PROJECT_PATH, "local")
+
+# local dataset configurations etc.
+LOC_PATH = environ.get('BYC_LOCAL_CONF', path.join(PROJECT_PATH, "local"))
 
 REQUEST_PATH_ROOT = "beacon"
-
-if "services" in LOC_PATH or "byconaut" in LOC_PATH:
+if "services" in PROJECT_PATH:
     REQUEST_PATH_ROOT = "services"
 
 #------------------------------------------------------------------------------#
 # Database settings
 #------------------------------------------------------------------------------#
 
-DB_MONGOHOST = environ.get("BYCON_MONGO_HOST", "localhost")
-
-# TODO: wrap them into object to make them mutable for local changes
-# or through environment variables like the host
-
-HOUSEKEEPING_DB = "_byconHousekeepingDB"
-HOUSEKEEPING_INFO_COLL = "beaconinfo"
-HOUSEKEEPING_HO_COLL = "querybuffer"
-
-SERVICES_DB = "_byconServicesDB"
-GENES_COLL = "genes"
-GEOLOCS_COLL = "geolocs"
+BYC_DBS = {
+  "mongodb_host": environ.get("BYCON_MONGO_HOST", "localhost"),
+  "housekeeping_db": "_byconHousekeepingDB",
+  "services_db": "_byconServicesDB",
+  "info_coll": "beaconinfo",
+  "handover_coll": "querybuffer",
+  "genes_coll": "genes",
+  "geolocs_coll": "geolocs",
+  "genomicVariant_coll": "variants",
+  "biosample_coll": "biosamples",
+  "individual_coll": "individuals",
+  "analysis_coll": "analyses",
+  "run_coll": "analyses",
+  "phenopacket_coll": "individuals",
+  "filteringTerm_coll": "collations",
+  "cohort_coll": "collations",
+  "publication_coll": "publications"
+}
 
 MONGO_DISTINCT_STORAGE_LIMIT = 300000
 VARIANTS_RESPONSE_LIMIT = 300000
@@ -69,42 +77,47 @@ BYC = {
 
   "BYC_DATASET_IDS": [],
 
-  "beacon_defaults": {
-    "defaults": {
-      "default_dataset_id": "examplez",
-      "test_domains": ["localhost"]
-    },
-  },
-
+  "default_dataset_id": "examplez",
+  "test_domains": ["localhost"],
+ 
   # ..._mappings / ..._definitions are generated from YAML files & should stay static
 
-  "argument_definitions": {"$defs":{}},
+  "argument_definitions": {},
+  "authorizations": {},
   "dataset_definitions": {},
   "datatable_mappings": {},
-  "entity_defaults": {"info":{}},
-  "filter_definitions": {"$defs":{}},
+  "entity_defaults": {},
+  "env_paths": {},
+  "filter_definitions": {},
   "handover_definitions": {},
   "interval_definitions": {},
+  "plot_defaults": {},
+  "request_meta": {},
+  "service_config": {},
   "test_queries": {},
   "variant_request_definitions": {},
   "variant_type_definitions": {},
 
   "loc_mod_pars": [
-    "argument_definitions",
     "authorizations",
-    "dataset_definitions",
-    "filter_definitions",
     "env_paths",
+    "filter_definitions",
     "datatable_mappings",
     "test_queries",
     "plot_defaults"
   ],
+  "info_responses": [
+    "beaconInfoResponse" ,
+    "beaconConfigurationResponse",
+    "beaconMapResponse",
+    "beaconEntryTypesResponse"
+  ],
+  "data_responses": [
+    "beaconCollectionsResponse" ,
+    "beaconResultsetsResponse",
+    "beaconFilteringTermsResponse"
+  ],
 
-  "authorizations": {},
-  "env_paths": {},
-  "plot_defaults": {},
-  "request_meta": {},
-  "service_config": {},
 
   # -------------------------------------------------------------------------- #
 
@@ -113,7 +126,6 @@ BYC = {
   "response_entity_id": None,
   "response_entity": {},
   "response_schema": "beaconInfoResponse",
-  "bycon_response_class": "BeaconInfoResponse",
   "returned_granularity": "boolean"
 }
 
@@ -159,8 +171,7 @@ BYC_UPPER = [
   "EFO",
   "UBERON",
   "NCIT",
-  "PATO",
-  "pubmed"
+  "PATO"
 ]
 
 ################################################################################